updated vasp convergence workflows
